Extract lexer helpers
