+++ src/parser/mod.rs
@@ -12,195 +12,14 @@ use rowan::{GreenNode, GreenNodeBuilder, Language};
 
 use crate::{DdlogLanguage, Span, SyntaxKind, tokenize};
 
-/// Convert a punctuation [`SyntaxKind`] to a printable character.
-///
-/// Diagnostic messages use symbolic token kinds which are not human
-/// readable. This helper maps those kinds to the punctuation displayed in
-/// errors.
-///
-/// # Examples
-///
-/// ```rust,no_run
-/// use ddlint::{parser::token_display, SyntaxKind};
-/// use chumsky::error::Simple;
-///
-/// let err = Simple::custom(0..0, format!(
-///     "expected '{}'", token_display(SyntaxKind::T_RPAREN)
-/// ));
-/// ```
-fn token_display(kind: SyntaxKind) -> &'static str {
-    match kind {
-        SyntaxKind::T_LPAREN => "(",
-        SyntaxKind::T_RPAREN => ")",
-        SyntaxKind::T_LBRACE => "{",
-        SyntaxKind::T_RBRACE => "}",
-        SyntaxKind::T_LBRACKET => "[",
-        SyntaxKind::T_RBRACKET => "]",
-        SyntaxKind::T_COMMA => ",",
-        SyntaxKind::T_COLON => ":",
-        SyntaxKind::T_SEMI => ";",
-        _ => "",
-    }
-}
+#[macro_use]
+mod lexer_helpers;
+use lexer_helpers::{atom, balanced_block, balanced_block_nonempty, ident, inline_ws};
 
 mod token_stream;
 
 mod span_collector;
 use span_collector::SpanCollector;
-
-/// Iterate over the token stream and dispatch handlers by [`SyntaxKind`].
-///
-/// The macro expects a parsing context `ctx` with a `stream` field. It loops
-/// until the stream is exhausted, invoking the handler associated with each
-/// recognised kind. Handlers must advance the stream to consume the tokens they
-/// process. Any unhandled kind is skipped.
-///
-/// # Examples
-///
-///
-/// ```
-/// struct State<'a> {
-///     stream: TokenStream<'a>,
-/// }
-///
-/// fn handle_kw(st: &mut State<'_>, _span: Span) {
-///     st.stream.advance();
-/// }
-///
-/// let mut st = State { stream: TokenStream::new(&tokens, src) };
-/// token_dispatch!(st, {
-///     SyntaxKind::K_IMPORT => handle_kw,
-/// });
-/// ```
-macro_rules! token_dispatch {
-    ( $ctx:ident, {
-        $( $kind:path => $handler:ident ),* $(,)?
-    } ) => {{
-        while let Some(&(kind, ref span_ref)) = $ctx.stream.peek() {
-            let span = span_ref.clone();
-            match kind {
-                $( $kind => $handler(&mut $ctx, span.clone()), )*
-                _ => $ctx.stream.advance(),
-            }
-        }
-    }};
-}
-
-/// Parser recognising whitespace and comment tokens.
-///
-/// The combinator is useful for padding other parsers where whitespace is
-/// allowed. It matches [`SyntaxKind::T_WHITESPACE`] and
-/// [`SyntaxKind::T_COMMENT`] tokens and discards them.
-///
-/// # Examples
-///
-/// ```no_run
-/// use ddlint::parser::inline_ws;
-/// let parser = inline_ws().repeated();
-/// ```
-fn inline_ws() -> impl Parser<SyntaxKind, (), Error = Simple<SyntaxKind>> + Clone {
-    filter(|kind: &SyntaxKind| matches!(kind, SyntaxKind::T_WHITESPACE | SyntaxKind::T_COMMENT))
-        .ignored()
-}
-
-/// Parser for an identifier padded by optional inline whitespace.
-///
-/// # Examples
-///
-/// ```no_run
-/// use ddlint::parser::ident;
-/// let parser = ident();
-/// ```
-fn ident() -> impl Parser<SyntaxKind, (), Error = Simple<SyntaxKind>> + Clone {
-    just(SyntaxKind::T_IDENT)
-        .ignored()
-        .padded_by(inline_ws().repeated())
-}
-
-/// Parser for a rule atom: `Ident` with optional argument list.
-///
-/// # Examples
-///
-/// ```no_run
-/// use ddlint::parser::atom;
-/// let parser = atom();
-/// ```
-fn atom() -> impl Parser<SyntaxKind, (), Error = Simple<SyntaxKind>> + Clone {
-    ident()
-        .clone()
-        .then(
-            just(SyntaxKind::T_LPAREN)
-                .padded_by(inline_ws().repeated())
-                .ignore_then(
-                    filter(|kind: &SyntaxKind| *kind != SyntaxKind::T_RPAREN)
-                        .ignored()
-                        .padded_by(inline_ws().repeated())
-                        .repeated(),
-                )
-                .then_ignore(just(SyntaxKind::T_RPAREN))
-                .or_not(),
-        )
-        .ignored()
-        .padded_by(inline_ws().repeated())
-}
-
-/// Parser for a balanced token block such as parentheses or braces.
-///
-/// The parser consumes the opening delimiter, then all tokens until the
-/// matching closing delimiter while tracking nested pairs. Whitespace and
-/// comments are permitted between tokens. An error is produced if a closing
-/// token appears without a corresponding opener.
-fn balanced_block_with_min(
-    open: SyntaxKind,
-    close: SyntaxKind,
-    min: usize,
-) -> impl Parser<SyntaxKind, (), Error = Simple<SyntaxKind>> + Clone {
-    use std::cell::Cell;
-
-    let depth = Cell::new(0usize);
-    just(open)
-        .padded_by(inline_ws().repeated())
-        .ignore_then(
-            filter_map(move |span, kind| match kind {
-                k if k == open => {
-                    depth.set(depth.get() + 1);
-                    Ok(())
-                }
-                k if k == close => {
-                    if depth.get() == 0 {
-                        Err(Simple::custom(
-                            span,
-                            format!("unexpected '{}'", token_display(close)),
-                        ))
-                    } else {
-                        depth.set(depth.get() - 1);
-                        Ok(())
-                    }
-                }
-                _ => Ok(()),
-            })
-            .padded_by(inline_ws().repeated())
-            .repeated()
-            .at_least(min),
-        )
-        .then_ignore(just(close))
-        .ignored()
-}
-
-fn balanced_block(
-    open: SyntaxKind,
-    close: SyntaxKind,
-) -> impl Parser<SyntaxKind, (), Error = Simple<SyntaxKind>> + Clone {
-    balanced_block_with_min(open, close, 0)
-}
-
-/// As [`balanced_block`] but requires at least one token inside the delimiters.
-fn balanced_block_nonempty(
-    open: SyntaxKind,
-    close: SyntaxKind,
-) -> impl Parser<SyntaxKind, (), Error = Simple<SyntaxKind>> + Clone {
-    balanced_block_with_min(open, close, 1)
-}
 
 /// Result of a parse operation.
 #[derive(Debug)]
